Implement Drag&Drop for Windows
Uses .NET Framework internals to implement Drag&Drop events
for the WindowsGameWindow. The events themselves are defined on
DesktopGameWindow so that other desktop platforms can use the same event
infrastructure later on by manually creating the necessary EventArgs.
Further, the old dragDrop and dragEnter methods in BaseGame have been
removed as well as old virtual On* methods that required WinForms for
their signature.