Fix default audio device tracking
The tracking of the default audio device was broken. When a change
in the default audio device was registered, the device switch
freed all existing Bass handles, causing existing AudioTrackBass and
AudioSampleBass instances to have invalid handles and thus behave
incorrectly. The fix involves not freeing Bass handles and instead
changing the device of existing channels and samples.