Rework Items of Dropdown menu to Key-Value-Pair
The items of dropdown menus are now represented as key-value-pairs.
This helps to separate the actual stored value the Bindable interacts
with from the displayed text by allowing users to specify any text
for any value. Effectively, the string key can be interpreted as the
user-facing text, paired up with the value that should be set when
the user selects the corresponding item. Intuitively, this might be
interpreted as a Dictionary, but actually using one here can cause
issues if a Dropdown Menu attempts to place the same string key
twice with different values, which would simply be impossible if
backed by a dictionary.