Cache DrawSize to improve performance
Since we already invalidate DrawInfo (which depends on DrawSize) at
the correct places, it follows that invalidating DrawSize whenever
DrawInfo is invalidated is necessarily sufficient (although there may
be opportunity to be more aggressive, i.e. invalidate it less often).

This change alone results in a three-fold performance increase within
the relative size benchmark, and almost a two-fold performance
increase in song select of the osu game.