Only update details of VisualisedDrawables when they are visible
This greatly improved performance once more, resulting in draw vis
being finally useful when at the highest-level draw node, even inside
song select.

It is accomplished in a slightly hacky way (it assumes no scaling
happens between the TreeContainer and the VisualisedDrawable).
This should be acceptable due to the WIP/debugger nature of the
draw vis.