Improve validation for BindableNumber<T>
BindableNumber unnecessarily checked the type argument whenever the
constructor was called, instead of using the type constructor
to validate the correctness of the type argument. Furthermore the
validation check would yield a false positive if a custom implementation
of IConvertible reports an allowed TypeCode, because that custom
implementation cannot be one of the built-in numeric types and
both the Set<U> and Add<U> method rely on T being one of
the builtin numeric types.