Refactor FlowContainer
The design of FlowContainer has been drastically revised. Instead of
always aligning elements in a grid, the new design uses a new
IFlowStrategy interface to determine positioning of the contents of a
FlowContainer. This allows for much more flexibility and end-user
extensibility of FlowContainer, while remaining easy to use at its
core. The existing behaviour of FlowContainer has been moved to the
FillFlowStrategy class and is used as the default behaviour of the
FlowContainer.