Refactor SpriteText based off FlowContainer
The changes to FlowContainer imply that Spacing and similar properties
of Text can no longer be controlled directly. Since SpriteText should
eventually get its own flow strategy and not allow the user to use a
custom one, all necessary properties to control text like it used to be
controlled from the outside have been exposed specifically, with better
names to signify their meaning in the context of text.