Fix incorrect drawable being highlighted by DrawVisualiser
Sometimes, the highlighted drawable isn't the one hovered in the
DrawVisualiser. Reason for that is slightly unstable ordering of
hover calls and sporadic unhover of the DrawVisualiser itself.
This commit addresses these problems fully by sorting the currently
hovered VisualisedDrawables by their nesting depth and always
highlighting the target of the VisualisedDrawable with highest
nesting depth.