Add rudimentary support for rigid body simulation
This commit introduces RigidBodySimulation, which allows performing
rigid body simulation on the children of a given container. For now,
the simulation is limited to impulse responses and does not support
resting contact well. Additionally, only centre origin, absolute
positioned children are permitted. Performance optimizations are
also an area of future work.