Do not lose focus when commiting in TextBox
When a user hits Enter to commit the contents of a TextBox, this used
to trigger a loss of focus for the TextBox, and then OnFocusLost used
to invoke the OnCommit-Handler on the condition that the enter key
is held. Now, hitting the enter key directly invokes OnCommit if the
TextBox is currently focused and does not remove focus from the
TextBox.