Quick fix for Raw Input cursor leaving the osu!lazer window.
It appears the OpenTK MouseState for Raw Input does not map the
coordinates received to the window, so a quick check against the bounds
of the current window allows for proper readjustment. This fix does not
work on windowed osu!lazer, but that is beneficial as the cursor
seamlessly follow the OS cursor.