Quick fix for Raw Input cursor leaving the window
It appears the OpenTK MouseState for Raw Input does not map the
coordinates received to the window, so a quick check against the bounds
of the current window allows for proper readjustment. This fix does not
work on windowed programs, but that is beneficial as the cursor
seamlessly follows the OS cursor.