Get rid of custom lifetimelist and directly allow implementing Compare
Custom lifetimelists expose unnecessary API surface to the user.
Since this API endpoint was not used and we plan to remove
lifetimelist altogether, the parameter was removed.
This also allows us to simplify the sorting API of children of
CompositeDrawable, now requiring only to override a Compare
method rather than providing a whole custom IComparer<Drawable>.