Default frame limiter to a refresh rate of 60 Hz
In some cases OpenTK provides the framework with an invalid refresh rate (e.g. 0 Hz). In such cases the frame limiter was previously broken; it did nothing. This commit instead defaults the frame limiter to using a refresh rate of 60 Hz (and multiples thereof) as it is most likely the user has a 60 Hz display. If the user has a higher refresh rate, then the 8x or unlimited frame limiters should still provide sufficient functionality for good use.