Allow children to be loaded in containers which are still in a Loading state
Conside the case where a Container has an autosize content with a drawable inside it. The drawable won't be considered for autosize until after the first UpdateSubTree without this change, as UpdateChildrenLife's (when called in load()) checkChildLife will return early.

Therefore, querying DrawSize in LoadComplete() will fail without this change, where it is expected to work.