Fix corner-case with named loggers
When two named loggers had the same name but cased differently
they would target the same file, but generate two instances of the
Logger class, which could lead to contention on the log-file in
multi-threading scenarios. Loggers are now always identified by their
name in lower-case to avoid this scenario.