Add a timeout for thread exiting
Under certain conditions, the audio thread is getting stuck during exit due to an underlying bass call. This should hopefully resolve that issue.

On top of this, we *do* always want to have a failsafe timeout for all threads to the game itself eventually gives up and leaves thread cleanup to the OS.