Fix broken texture coordinates in PathDrawNode/CircularProgressDrawNode
This wound up being a recursive spiral of fail. The pattern
`1 / Texture.{Width,Height}` is broken because those properties are
integers, and  everything rounds to zero. Except for 1x1 textures where
it doesn't round. But this whole code was broken for texture atlases
because it was computing coordinates without taking the atlas into
account, assuming it was a single texture. But it happened to work for
WhitePixel because 1 - 1 = 0 and so the bottom right coord moved (by
what the code thought was one pixel, but in reality was the whole atlas)
to the top left corner, which is where the white pixel is.

Blergh.

* closes (potentially) ppy/osu#3215