Fix game host overwriting culture set in tests
A test failure was spotted while locally running tests added in #2895
for setting culture loaded from the framework config, pertaining
specifically to the invariant culture test. The desired invariant
culture was being silently overwritten with the system culture.

The cause of this scenario was detected to be related to the way
headless tests are ran.

1. When GameHost starts, it loads in default settings, as specified by
   GetFrameworkConfigDefaults(), and merges them with framework settings
   read from the framework.ini file. The invariant culture is used as
   the default locale, so at this point the bindable associated with the
   locale setting has an empty string value.

2. In accordance with the BindValueChanged callback set in GameHost,
   all threads get their culture set to the invariant one.

3. When ThreadCultureTest is supposed to be run, TestScene performs
   setup logic to prepare the test for execution in SetUpTestForNUnit().
   In particular, it uses the NUnit test execution context to set the
   culture of all running threads to what NUnit says. This silently
   overwrites changes made in step (2), without changing the locale
   bindable value.

4. In the TestDefaultCultureIsInvariant test itself, the locale setting
   is set to the empty string again. If the value was set to the default
   in step (1), this means that the ValueChanged delegate will not fire
   due to the values being equal, and therefore the thread culture will
   not be set to what the test specifies.

To fix the above scenario, force setCulture() to explicitly trigger
change events after setting the locale in the test.

It's very possible this issue has not surfaced thus far due to test
execution order - if another test ran before, it might have changed the
locale setting value between steps (1) and (4), making the scenario
not reproducible.