Replace normal list with concurrent collection
A sporadic test failure has been revealed in a CI build for PR #2915.
After inspecting the failing test the problem was determined to be
caused by scheduled adds to a non-thread-safe List<T>, which caused
waits for all three desired items to be added to it to time out. This
race condition could be reproduced by repeating any test in
ThreadCultureTest enough times.

Replace the List<T> with a ConcurrentBag<T> to prevent race conditions
from occurring. The change was tested by repeating all tests in the
ThreadCultureTest fixture 500 times - the failure could not be
reproduced after introducing it.