Do not close input if it wasn't opened
If avformat_open_input() failed when trying initialising a decoder,
but a format context was created, it was possible that during disposal
an attempt to call avformat_close_input() without a matching open led
to a segfault.

Prevent unmatched close calls by introducing a flag tracking if an input
was ever opened.