Prevent first sprite from overlapping white pixel
Due to recent changes in the texture atlas area, a scenario could arise
in which the first sprite to be added on a page of the texture atlas
would receive a position that partially overlaps the white pixel in the
upper right corner of the atlas.

The regression was caused by no longer using the Add() method to add the
white pixel to the atlas, but by manually creating the upload and
sub-texture. Due to this, inside of Reset() currentPosition would first
be set to (0, 0), and then to (PADDING, PADDING). The previous way,
using Add(), would silently mutate the value of currentPosition and
therefore work fine.

To resolve, manually set the current position in the atlas to ensure
space for the white pixel on the X axis and additionally the necessary
padding on both X and Y axes.