Workaround a BASS issue with parsing WAV files
BASS has an issue parsing WAV files with the "data" section size being
zero. When this is the case, it uses the entire rest of the file as
audio data, making a very unpleasant sound for files that are supposed
to be silent. This commit solves this by parsing the file and detecting
this case, and passing no data to BASS when this happens.