Don't log performance issues on non-async lookups if they are cached
While this is still generally a no-go (you can never be sure that the
underlying texture/sample lookups are cached, which is why we always
logged) the overhead from logging these issues can become overwhelming,
with megabytes of output per second.

This should still be enough to generally cover cases a developer would
care about.