Add bounds check to avoid storing borderless position to config
This handles a weird case where RESIZE events are not fired in
borderless or exclusive fullscreen on windows, instead being delayed
until the next time WindowState is set to Normal.

This was causing a position of -1,-1 to be stored to the config (coming
from the windows implementation of SetBorderless). Rather than
explicitly handling the -1 case, doing a bounds check feels more "sane".