Move GL disposal to class with improved safety
The previous logic was prone to rare and subtle races between the
enqueue and execution operations. As the two are usually called from
different threads, the execution operation could potentially observe the
list in an inconsistent state.

The race occurred when thread A was enqueueing a disposal when the queue
was empty, while thread B was trying to execute enqueued operations. Due
to the lack of thread safety of `List<T>`, thread B could potentially
read a null `PendingDisposal` and crash with an exception.

To alleviate this lack of thread safety, a new scheme is implemented,
using two lists. Enqueue operations only add to a `newDisposals` list
under a lock. Before executing disposals, the same lock is taken to move
items from `newDisposals` to the end of `pendingDisposals`. Presuming
that disposal execution will never be called in a non-thread-safe
manner, this ensures execution correctness.