Stop handling unobserved exceptions
We have generally agreed that these should be left unobserved, as it is
quite common to fire a `Task.Run` without expecting it to cause the
whole application to crash.

I've left logging in as that part will definitely be useful for tracking
down edge cases where we actually do care about the exceptions.