Revert weird CursorContainer logic that causes incorrect cursor positioning
This logic would result in a displayed cursor drifting from the true
cursor position if the `Container` (or any parent hierarchy) is
transformed in any way.

Expected behaviour is that no matter how the draw hierarchy is
transformed, the cursor is always displayed at the tracking location
(based on the parent `InputManager`'s `Mouse` position.