Fix assertion tripping when seeking from within the mixtime callback
The recent changes to `TrackBass` added an instantaneous current time
value update on a seek. As it turns out, it is not necessarily safe to
unconditionally run that update, as BASS callbacks (such as the mixtime
callback) could run from a separate unmanaged thread rather than the
audio thread, tripping the `CanPerformInline` assertion check if a track
had `Looping` and `RestartPoint` set.

To fix, add an explicit check for whether the current time update can be
safely performed via `seekInternal()` before trying to attempt it.