Fix culture changes not working when single-threaded
As revealed by efforts to add framework test runs in single-threaded
mode as part of CI, the culture-switching code would not work if the
game was started single-threaded without ever switching to
multi-threaded. The cause of this was two-fold:

* First off, when the game is started in single-threaded mode, the four
  `GameThread`s are never actually created via `ThreadRunner`.
  Therefore, the `GameThread.CurrentCulture` sets did nothing but set
  the culture value on the `GameThread`s. They did not actually change
  any thread's culture.

* Secondly, in single-threaded mode, the main thread is actually the one
  that is running frames of all four types (audio/draw/input/update) -
  but there was no actual place where the main thread's culture was
  changed, partly due to the above.

To resolve, extract a `SetCulture(CultureInfo)` method on `ThreadRunner`
and explicitly set the culture of that method's caller thread to the
desired culture, as well. The method does not branch depending on the
current execution mode to avoid desynchronisation of culture values
across the two available modes.