Fix SampleChannels with zero frequency not playing
This resulted in the playChannel() callback temporarily setting
`Playing` to false by virtue of setting `enqueuedPlaybackStart` to
false.

In multi-threaded execution, this is not a problem because `UpdateState`
will then run and set `Playing` back to true by setting `playing` to true.

In single-threaded execution, this is a problem because `UpdateState`
runs on the _next_ audio frame, at which point the `SampleChannel` has
`IsAlive = false` (due to `Playing = false`) and won't receive the
`UpdateState`.