Avoid changing windows time period when focus changes
In an attempt to figure out what is causing
https://github.com/ppy/osu/issues/12702 we looked at all the operations
we are performing on window active state change. This is the one that
stood out the most.

Since this is a hard one to reproduce (does not affect all users), just
going to get this out for user testing. Thinking about it, this could
only be affecting users that have no other applications system-wide
requesting a time period elevation. In general this is pretty rare as
even Chrome will request it.

For reference, we don't do this in stable, and there is little reason to
reset it on active state change (although it may be something we would
consider if the game is ever minimising to system tray and expected to
use zero resources).