Use alternative method of fitting popover contents inside container
The previous method, which moved the body of the popover inside of it,
was causing one-frame issues due to how
`BoundingBox`/`RequiredParentSizeToFit` work. In short, moving the body
led the popover's bounding box to be wider/taller than its actual
contents are, because the parent was trying to reserve space for the X/Y
position on top of the actual body content size.

To counteract this, move the popover as a whole in the same way, but
move the arrow back the other way. That no longer affects `BoundingBox`
in that same manner.