Fix directory/file selector potentially entering recursion loop
It was possible for a directory change to cause an endless recursion. If
both the new value and the old value of `CurrentPath` was inaccessible
to the user (due to being deleted, rendered inaccessible in another
manner, etc.), then the selector would first try to enter the new
directory, then failing that, endlessly attempt to enter the old
directory.

To resolve, iteratively traverse the filesystem graph upwards to find an
accessible directory on failure, and only set `CurrentPath` once when a
suitable directory has been found.

Co-authored-by: Pasi4K5 <pascal.pomper03@gmail.com>