Remove concept of passthrough mixer
I thought this was a nice idea - if the component is removed from the
mixer (via mixer.Remove(component)), then it would get a passthrough
mixer that defaults to the basic BASS commands we've been using until
now.

In practice, this doesn't really work as expected because we force all
channels to be decode streams in order to work with BassMix.
Theoretically it should be possible to unset this flag, however doing so
is non-trivial because we also define sample channels as streams.

The simplest way is to say we have a global mixer