Remove AudioMixerOverlay
Maybe temporarily, maybe permanently. As more audio mixers are applied,
we'd need ways to identify them in the hierarchy as well as their
contained channels. This could be done by say, the sample name, however
SampleChannel (or Track) doesn't know the resource name at this point in
time.

If anything happens, this should be implemented as a follow-up taking
the above into consideration.