Rename `Drawable.Click` to `Drawable.TriggerClick` to avoid naming conflict
This is primarily to avoid a naming conflict with
`ManualInputManager.Click`, but also better aligns with the fact that
it's triggering an event (via `TriggerEvent`) internally.

For additional safety, `ManualInputManager` will also throw on usage of
this method (as it is a common gotcha mistake when trying to create
tests). I haven't bothered with the obsoleted method here since the user
should see an obsolescence notice in such cases.