Remove `Transform` construction-time delegate overhead
This restructures how completion and abortion events are propagated from `Transform`s to `TransformSequence`s to avoid delegate allocation overheads (due to self-capture).

Note that this is a "breaking change" in that it makes `Transform.OnComplete` and `Transform.OnAbort` `internal`, but using these publicly would have broken things regardless so this should be considered a code safety improvement.

Will add more xmldoc if this is seen as the correct path / structure.