Update `getGlyphHeight` to take post-scaling glyph size instead
Also renames the field to become `useFullGlyphHeight` to better reflect
what it actually does here, also matches with `SpriteText.UseFullGlyphHeight`.

If I understand this properly, when setting this to false, it should
take the glyph's texture height instead so that text get properly
centered, but for some reason it's still including `YOffset` which could
leave an empty area at the top of the text.