Update `AddNewLine` to insert post-scaling glyph height on empty line
This one required a bit more work as CSS usually takes the "glyph scale"
of the font specified in the `font-family` as far as my manual testing went.

I think it should be fine to create a `GetFont` method returning
`IGlyphStore`s for that purpose, as it's also required in visual testing
to visualise the baseline position in a sprite text.