Avoid throwing errors on failure to flush stream
Also saw this occur on macOS in the form of:

```csharp

System.ObjectDisposedException : Cannot access a closed file.
   at System.IO.FileStream.Flush(Boolean flushToDisk)
   at osu.Framework.Platform.FlushingStream.Dispose(Boolean disposing)
   at System.IO.Stream.Close()
   at osu.Game.Tournament.Models.StableInfo.SaveChanges() in
/Users/dean/Projects/osu/osu.Game.Tournament/Models/StableInfo.cs:line
58
```

See
https://github.com/ppy/osu/blob/879b1ab046eefd501e82a8d4d4ce4baa124872ac/osu.Game.Tournament/Models/StableInfo.cs#L49,
which uses a `StreamWriter` without specifying `leaveOpen=true`. This
may be a point of concern/investigation though (ie. does this leave a
case where the resulting file could be NUL byte filled?). I seem to
recall `StreamWriter` performing its own flush operations, but possible
not with the `true` parameter we added.

Either way, for the other case reported in the issue this try-catch is
likely still what we want.

Closes https://github.com/ppy/osu-framework/issues/4779.