Fix audio device 0 re-init on linux
We explicitly don't free device 0 on Linux in order to prevent BASS
native deadlocks.

Thus, re-initting device 0 will always result in
`Bass.LastError = Errors.Already`, a "fail" condition where the correct
device is chosen but the returned `didInit` value is always false, such
that `AudioComponent.UpdateDevice()` is never called on the hierarchy
for the device and components remain on a previous device.

The fix is simple - return `true` for this particular case on Linux. The
rest is a bit of shuffling code around to more natural places.