Remove text update scheduler entirely
It was added in 2016, nobody can remember why it existed anyway, IME
paths that could be the cause for it to be added are not hooked up
anyway, and key event handlers happily ignored that scheduler entirely
and just mutated `text` to their heart's content.

In the meanwhile anything setting `Text` in conjunction with custom
validation rules (like max length, etc.) would experience weird bugs
due to the `Text` setter going through the scheduler and therefore
having a delayed effect. c2350c2 removed that scheduler's usage in the
aforementioned setter.