Add screen stack logging
Generally pushing screens is quite a rare event, and should add context
to most log usages (tracking down issues reported by users etc.).

We already did this osu! side but moving to framework seems to make
sense. I've left at a `Verbose` level so it is actually logged in
release builds.

Not sure about the log format, especially the `[]` as they are also used
for the verbosity level.