Fix async delay oversight in `TestExitBeforePush`
This test was checking that a cancelled load screen (aka `ValidForPush` being `false`)
will not be pushed or `IsCurrent` after it finishes loading. Due to the
lack of an `Until` step, it's feasible the resume to previous screen was
not done in time for the assert.

As seen in
https://github.com/ppy/osu-framework/runs/4614508910?check_suite_focus=true
(first failure).