Fix `LoadComponentAsync` running load continuation asynchronously
There is no reason for this to be run asynchronously. It's a lean
continuation and has no user code. This should improve performance of
async load tasks as there is one less thread transition.

It also means if the TPL thread pool is saturated, drawables will still
load. This may in turn cause unblocking of other threads in the thread
pool (due to bad user implementations in drawable load code, for
instance) and avoid potential deadlocks.