Consume touch `IsActive` state rather than checking for "any" touch for mouse mapping
While the previous method worked fine, it means that a consumer would
not be able to alter the flow of events (ie. blocking the "click"
operations by changing the incoming touch).

In osu!, we wish to use this to allow the user to block the "click"
portion while still convey movement of the "cursor".