Add support for Android pointer capture
Equivalent to desktop relative mode.
The behaviour and code is modelled after the desktop `MouseHandler`.

Enabled by default so that the cursor can be hidden if that was requested
trough `IWindow.CursorState`.

All mouse events changed to use `GetX()` and `GetY()` as that is more correct
and has the same behaviour as `GetHistoricalX()` etc.