Remove `MemoryStream` shortcutting
It's too much effort to discern whether the `GetBuffer` call will
succeed or not. The newer alternative `TryGetBuffer` returns an
`ArraySegment`, so if we want to return as a simple `byte[]` would
require a `ToArray` call, which is undoing any optimisation that we
would gain from this shortcut in the first place.