Use higher-precision attributes to compute border gradient
On non-desktop GL ES 2.0 platforms (such as Android and iOS)
`v_TexCoord` and `v_TexRect` attributes (which are declared as
`mediump`) have insufficient precision to properly bilinearly
interpolate the colour of the border gradient for each fragment, which
leads to the gradient degenerating to a single colour.

Use `v_MaskingPosition` and `g_MaskingRect` instead, both of which are
`highp`.