Fix crash in `SafeAreaContainer` when working with negative sizing
I'm not 100% sure this will provide the correct behaviour when
negatively sized. It's hard to tell what the expectation would be in
this case (sizing the `SafeAreaContainer` in this manner will give
completely unexpected results unless it completely ignores the negative
component... since the intention of these containers is to handle device
edge scenarios).

Fixes https://github.com/ppy/osu-framework/issues/5015.