Prevent FillFlowContainer elements jumping rows with Spacing transform
If Spacing was transformed with a Direction of FillDirection.Full, some
elements might "jump" between rows (see ppy/osu#9155) during the
transform.

This commit introduces an "intended horizontal spacing" to
FillFlowContainer, which is used instead of Spacing.X when to place
elements on a new row. Specifying this before a Spacing transform will
prevent items from jumping between rows during the transform.