Refactor and update sample playing state on `Stop` call
The refactoring ensures `playing` is set on `Play`/`Stop` instead of
`playChannel`/`stopChannel` (which has been renamed to
`playInternal`/`stopInternal` inline with track code). This is to ensure
the zero frequency handler logic doesn't affect the `playing` state,
since we want the channel to appear as playing but pause it internally.