Standardise `DrawableMenuItem` colour update logic
As of https://github.com/ppy/osu-framework/pull/5054, it was feasible
that `OnSelected` would be fired by a `DrawableDropdown` on its child
items before they were `IsLoaded`. If an implementation relied on
`OnSelected` to update the colour state correctly, this would have
caused a potential regression in behaviour due to calls to
`Update*Colour` not happening when `!IsLoaded`.

To fix this, I've standardised all calls to `UpdateBackgroundColour` and
`UpdateForegroundColour` to be scheduled and debounced. I've also
replaced multiple methods of performing the same colour set in different
ways.

The `AddOnce` was a large improvement, as each item was previously
having it called unnecessarily 3-5 times before initial display.

I've tested this against osu! manually. It's hard to test framework-side
because it was caused by a non-standard osu! side implementation of
colour/hover/select handling.