Reword comment in textbox's `PlatformAction.Paste` to better explain what's happening
Was hoping this can implicitly go away, but since we enable the paste functionality in software keyboard, the Cmd+V shortcut will still activate that next to being enqueued in `IOSHardwareKeyboardHandler`, potentially causing a double-paste.

There's nothing that can be done to block this shortcut at iOS side, so our only choice is to keep avoiding the action at `TextBox`.