Fix unsafe file writes in `RawCachingGlyphStore`
As noticed in https://github.com/ppy/osu-framework/issues/5177, writes
of these files have been unsafe until now if the file write is
interrupted partway. The length of the written file could be shorter
than expected.

This uses a newly added method to write to a temporary file before
moving to the final location.