Move `CreateFileSafely` implement to non-virtual to avoid potential misuse
Abstract methods in storage are only required where they are used to
change actual file or path access. Because `CreateFileSafely` is
performing all actions via calls to the `Storage` API, it does not need
an `abstract` implementation.

---

For the full ratification of this change:

While integrating these changes in osu! side, I found unexpected test
failures due to the fact that the `NativeStorage` implementation of this
function is passing itself as a storage.

In osu!, we have the concept of a `WrappedStorage`, where all calls for
path lookups are supposed to be re-routed. Due to the structure of how
this was implemented, it was actually impossible to get correct
behaviour, unless changing `WrappedStorage` to implement `NativeStorage`
(and use the existing implementation of `CreateFileSafely`, which would
then pass through the `WrappedStorage` correctly).