Rename "unlimited" to "basically unlimited" and add sane frame limiter
Games using osu!framework can generally run at *very* high
frame rates when not much is going on.

This can be counter-productive due to the induced allocation
and GPU overhead.
 - Allocation overhead can lead to excess garbage collection
 - GPU overhead can lead to unexpected pipeline blocking (and
   stutters as a result).
   Also, in general graphics card manufacturers do not test
   their hardware at insane frame rates and therefore drivers
   are not optimised to handle this kind of throughput.
 - We only harvest input at 1000hz, so running any higher has
   zero benefits.

We limit things to the same rate we poll input at, to keep
both gamers and their systems happy
and (more) stutter-free.