Suppress `SwapBuffers()` failures on android
For reasons that are currently unclear, sometimes `SwapBuffers()` calls
on android fail with `BAD_SURFACE` just before the game activity is
suspended, and fail just after the game activity is resumed - but if
given ample time to recover from the initial failures after resume, the
game will continue functioning as usual.

As a temporary solution, suppress `SwapBuffers()` failures. The ideal
solution would be to hook onto one of the many callbacks of
`GLSurfaceView`, but in testing there does not appear that the obvious
ones fit the bill.