Allow disabling forced GC during test runs
I'm not saying we should do this, but I want the flexibility to
experiment with it. Once we have things in a stable state, it may make
sense to disable for Github Actions to improve performance of test runs
(from recollection, about 40% faster) and enable only when an issue
comes up, or as a background runner on teamcity or otherwise.