Change `BindTexture` flow to avoid duplicated logic in renderer implementations
Previously, binding a texture would happen in the following order:
 1. `BindTexture(Texture)` (optimisation checks)
 2. `INativeTexture.Bind()` (upload, checks, increase bind count)
 3. `BindTexture(int)` (set wrap mode, store texture in array for checks).

`BindTexture(int)` contains logic that should be shared across all
renderer implementations (set wrap mode and store texture array),
but the signature of the method is GL-specific and it's called
from `INativeTexture.Bind()` which requires implementation by consumer.

This all turned into an awkward flow, so I've went ahead and refactored
the flow to:
 - have `INativeTexture.Bind()` not exist and merge it with
`BindTexture(Texture)`.
 - change `BindTexture(int)` into `BindTexture(INativeTexture)`.
 - add `SetTextureInternal` for renderer implementations to perform the
   calls to bind the texture with the graphics API (GL, etc.).