Fix unbind action caching occurring too late
This was causing osu! side failures. It "regressed" with #5411, but was
only "working" previously due to the relevant base types being cached
elsewhere. So a bit of luck that things have been working until now.

I'm not sure how to write test coverage for this (or if it's required),
but I've run osu! side tests and confirmed they now pass.