Fix VBO incorrectly flushing on no-op shader bind
Previously, we were maintaining a variable pointing to the current
shader such that if a `Pop` then `Push` happened resulting in the same
shader being bound, it would not trigger a buffer flush or any draw call
overhead.

This check was removed incorrectly.

Regressed in https://github.com/ppy/osu-framework/pull/5404.

Closes ppy/osu#20421.