Adjust assertions after search algorithm change
The test was previously checking that all the `HeaderContainer`s are not
present. This however breaks if one of them is caught early as
`CanBeShown = false`, as then the search algorithm will not descend into
that subtree at all, and *technically* leave some `HeaderContainer`s
present - but the ones we actually care about, i.e. the topmost-level
ones, will be hidden correctly.