Fix incorrect transferal order in `RangeConstrainedBindable.CopyTo()`
Regressed in 9faae09581d9c4aa24175ed78197720fd983e40e.

When copying over values of a range-constrained bindable, the range
bounds need to be transferred first, because assignment to `.Value` of a
range-constrained bindable will automatically clamp the value to
[`MinValue`, `MaxValue`], leading to a scenario where a new value being
transferred across would get clamped to stale bounds.

This affected both `.CopyTo()` and `.BindTo()`, as the latter calls the
former.