Fix search container filtering too much
As it turns out, 99915f9286d10b12b00479346c30005832ee8823 introduced a
performance regression in `SearchContainer`. Due to borrowing from #5154
too much, the search logic was changed in order to hook into
`InvalidateLayout()` to perform search, rather than just do it when a
new internal child was added, as was the case previously.

For the sake of completeness, let's review the possible causes for
`InvalidateLayout()` firing with respect to filtering:

- `AddInternal()`: Valid, as the new child needs to be examined against
  the current filter, and preserved by reverting to the old code.
- `RemoveInternal()`: Mostly irrelevant, as long as the presence of the
  removed drawable in the container doesn't impact the filter state of
  any of its siblings. Let's not support that case for now.
- `ClearInternal()`: Totally irrelevant, as anything that could be
  filtered is now gone.
- `SetLayoutPosition()`: Hopefully irrelevant.
- `UpdateChildrenLife()`: Not relevant. Nothing in `SearchContainer`
  considers lifetime explicitly.
- `FlowContainer.childLayout` becoming invalid: The invalidation flags
  there are `RequiredParentSizeToFit | Presence`. Neither should be
  relevant for filtering, given the direction taken with
  `IConditionalFilterable`.
- `Direction` and `Spacing`: 100% irrelevant.